CHECKPOINT (not approved): item-cost phase 9 implement r1 — living-docs folder, two frontend handoffs, drift batch
Discharges the phase-9 GOVERNING block P1–P22 (the last phase of v1).

Docs
- docs/domains/item_economics/{README,api,events,states}.md + the docs/README.md
  domain-map row (P1 — the folder the 23_documentation contract mandates).
- Two frontend handoffs (R19-1): configuration (13 setup routes) and operational
  (10 routes), covering all 23 endpoints the project shipped, plus the removals,
  the twelve-value status vocabulary and the flow narratives (P6).
- docs/deploy runbook for the column-drop deploy ordering (P19); the domain doc
  states the generic rule without naming a revision.
- architecture/46_serialization_local.md rewritten and 05_errors_local.md created
  (P20) — the standing serialization divergence and the no-code-field error
  contract with the leading-token identity rule.
- README batch: tables index +9 rows and two corrections (P5), items/ and tasks/
  README drift incl. the D-4 line (P9), the nine prefix-map rows sorted into
  alphabetical place (P11), routers/README.md's PUT /api/v1/tasks body table
  repaired and its false "Autogenerated" banner replaced (P15).
- Frontend mirrors (P12), Application_contracts step aggregates + item value/cost
  semantics (P13), the old reassigned-steps handoff's total_cost_minor row (P18).

Tests
- tests/unit/docs/test_item_economics_docs.py — C1, parents[4] anchor, verbatim
  intention literals (P2).
- tests/unit/docs/test_item_economics_handoff_accuracy.py — hand-written route,
  identity, status and envelope-key sets asserted against the shipped artifacts.
- tests/unit/routers/test_phase9_item_economics_route_mirror.py — C4's
  hand-written 23-row arbiter over routers/README.md and the router's role gates.
- tests/unit/services/queries/item_economics/test_phase9_committed_filter_structure.py
  — the compiled-statement filter arbiter over the three recorded sites (P3).

Code (P4's enumerated allow-list, nothing else)
- Eleven Mapped[float] → Mapped[Decimal] annotations on Numeric columns.
- 4B N3: the redundant `and not version.is_deleted` dropped at :39 and :49.
- be9dfe42a035's docstring Revises: line corrected (prose only; rule-7 exemption
  stated in the Review log).
- The downgrade proxy now also matches raw-SQL DROP TYPE.
- Shopify sync-item ids compared as a set, not an ordered list.
- The C5 fixture closes its task before the FIRST handler run.
- The two-line hand format edit in test_phase8_serializers.py.

Suite 2249 passed / 23 failed / 1 deselected = 2273 selected; the 23-failure set
is byte-identical to the phase-1 baseline. DB left at head c1d2e3f4a5b6.
Archgraph +1 node (decision-money-audience-admin-manager-only) → 175/260.

+++ app/beyo_manager/models/tables/item_economics/cost_model_term.py
@@ -1,4 +1,5 @@
 from datetime import datetime, timezone
+from decimal import Decimal
 
 from sqlalchemy import Boolean, CheckConstraint, DateTime, Enum as SAEnum, ForeignKey, Index, Integer, Numeric, String, text
 from sqlalchemy.orm import Mapped, mapped_column
@@ -19,7 +20,7 @@ class CostModelTerm(IdentityMixin, Base):
     cost_model_version_id: Mapped[str] = mapped_column(String(64), ForeignKey("cost_model_versions.client_id", ondelete="RESTRICT"), nullable=False, index=True)
     name: Mapped[str] = mapped_column(String(100), nullable=False)
     calculation_type: Mapped[CostModelTermCalculationTypeEnum] = mapped_column(SAEnum(CostModelTermCalculationTypeEnum, name="cost_model_term_calculation_type_enum", create_type=True), nullable=False)
-    percent_value: Mapped[float | None] = mapped_column(Numeric(6, 3), nullable=True)
+    percent_value: Mapped[Decimal | None] = mapped_column(Numeric(6, 3), nullable=True)
     fixed_amount_minor: Mapped[int | None] = mapped_column(Integer, nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False, default=lambda: datetime.now(timezone.utc))
     created_by_id: Mapped[str] = mapped_column(String(64), ForeignKey("users.client_id", ondelete="RESTRICT"), nullable=False, index=True)
