+++ app/beyo_manager/models/tables/item_economics/item_cost_evaluation_term.py
@@ -1,4 +1,5 @@
 from datetime import datetime, timezone
+from decimal import Decimal
 
 from sqlalchemy import DateTime, Enum as SAEnum, ForeignKey, Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column
@@ -19,7 +20,7 @@ class ItemCostEvaluationTerm(IdentityMixin, Base):
     evaluation_id: Mapped[str] = mapped_column(String(64), ForeignKey("item_cost_evaluations.client_id", ondelete="RESTRICT"), nullable=False, index=True)
     name: Mapped[str] = mapped_column(String(100), nullable=False)
     calculation_type: Mapped[CostModelTermCalculationTypeEnum] = mapped_column(SAEnum(CostModelTermCalculationTypeEnum, name="cost_model_term_calculation_type_enum", create_type=False), nullable=False)
-    percent_value: Mapped[float | None] = mapped_column(Numeric(6, 3), nullable=True)
+    percent_value: Mapped[Decimal | None] = mapped_column(Numeric(6, 3), nullable=True)
     fixed_amount_minor: Mapped[int | None] = mapped_column(Integer, nullable=True)
     amount_minor: Mapped[int] = mapped_column(Integer, nullable=False)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False, default=lambda: datetime.now(timezone.utc))
