+++ app/beyo_manager/models/tables/item_economics/production_cost_basis_version.py
@@ -1,4 +1,5 @@
 from datetime import date, datetime, timezone
+from decimal import Decimal
 
 from sqlalchemy import Boolean, CheckConstraint, Date, DateTime, Enum as SAEnum, ForeignKey, Index, Integer, Numeric, String, text
 from sqlalchemy.orm import Mapped, mapped_column
@@ -21,9 +22,9 @@ class ProductionCostBasisVersion(IdentityMixin, Base):
     effective_to: Mapped[date | None] = mapped_column(Date, nullable=True)
     fixed_monthly_cost_minor: Mapped[int] = mapped_column(Integer, nullable=False)
     currency: Mapped[ItemCurrencyEnum] = mapped_column(SAEnum(ItemCurrencyEnum, name="production_cost_basis_version_currency_enum", create_type=True), nullable=False)
-    monthly_paid_hours: Mapped[float] = mapped_column(Numeric(8, 2), nullable=False)
-    planning_utilization_percent: Mapped[float] = mapped_column(Numeric(5, 2), nullable=False)
-    cost_per_worker_minute_minor: Mapped[float] = mapped_column(Numeric(12, 4), nullable=False)
+    monthly_paid_hours: Mapped[Decimal] = mapped_column(Numeric(8, 2), nullable=False)
+    planning_utilization_percent: Mapped[Decimal] = mapped_column(Numeric(5, 2), nullable=False)
+    cost_per_worker_minute_minor: Mapped[Decimal] = mapped_column(Numeric(12, 4), nullable=False)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False, default=lambda: datetime.now(timezone.utc))
     created_by_id: Mapped[str] = mapped_column(String(64), ForeignKey("users.client_id", ondelete="RESTRICT"), nullable=False, index=True)
     updated_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True, onupdate=lambda: datetime.now(timezone.utc))
